Trait libtls::error::LastError

pub trait LastError {
    fn last_error(&self) -> Result<String>;

    fn to_error<T>(errstr: String) -> Result<T> { ... }
}

Returns the last API error.

The Tls and TlsConfig structs both provide a way to return the last error as a String from the underlying API.

Required methods

fn last_error(&self) -> Result<String>

Return the last error of the underlying API.

The last_error method returns an error if no error occurred at all, or if memory allocation failed while trying to assemble the string describing the most recent error related to config.

Loading content...
Provided methods

fn to_error<T>(errstr: String) -> Result<T>

Returns the error string as an error object.

Loading content...